+++ IA/src/ParseArgs.py
@@ -6,21 +6,24 @@ class ParseArgs:
         self.host = 'localhost'
         self.port = -1
         self.name = ''
+        self.id = -1
 
     def check_type(self):
         try:
             self.port = int(self.port)
+            self.id = int(self.id)
         except ValueError:
             self.print_invalid_argument()
             sys.exit(84)
         
-        if self.port == -1 or self.name == '':
+        if self.port == -1 or self.name == '' or self.id == -1:  # Check if ID is provided
             self.print_usage()
             sys.exit(84)
 
         self.check_host()
         self.check_port()
         self.check_name()
+        self.check_id()
 
     def check_host(self):
         try:
@@ -39,6 +42,11 @@ class ParseArgs:
             print("Error: Name too long")
             sys.exit(84)
 
+    def check_id(self):
+        if self.id < 0:
+            print("Error: ID must be a non-negative integer")
+            sys.exit(84)
+
     def parse(self, args):
         for i in range(0, len(args), 2):
             if args[i] == '-p':
@@ -47,15 +55,17 @@ class ParseArgs:
                 self.name = args[i + 1]
             elif args[i] == '-h':
                 self.host = args[i + 1]
+            elif args[i] == '-id':
+                self.id = args[i + 1]
             else:
                 self.print_invalid_argument()
                 sys.exit(84)
 
         self.check_type()
-        return self.host, self.port, self.name
+        return self.host, self.port, self.name, self.id
 
     def print_usage(self):
-        print("USAGE: ./zappy_ai -p port -n name -h machine")
+        print("USAGE: ./zappy_ai -p port -n name -h machine -id id")
         sys.exit(84)
 
     def print_invalid_argument(self):
